deconstructed signup and login pages with the signup form and login form components
