+++ client/src/App.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
 import Navbar from "./components/Navbar";
-import Signup from "./components/Signup";
-import Login from "./components/Login";
-import About from "./components/About";
+import Signup from "./pages/Signup";
+import Login from "./pages/Login";
+import About from "./components/ImageCard";
 import Create from "./components/Create";
-import Search from "./components/Search";
+import Search from "./components/SearchForm";
 import "./App.css";
 
 class App extends Component {
@@ -12,11 +12,11 @@ class App extends Component {
     return (
       <div className="App">
         <Navbar />
-        {/* <Signup /> */}
+        <Signup />
         {/* <Login /> */}
         {/* <About /> */}
         {/* <Create /> */}
-        <Search />
+        {/* <Search /> */}
       </div>
     );
   }
